refactor(headless): split RSI replay boundaries
